+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
 ## future
-- Remove serde as default feature.
+### New features
 - Add debug implementation to majority of structs.
+- Add option to get removed callback functions with `Sink::take_callback` and
+  `Sink::take_err_callback`.
+
+### Breaking changes
+- Remove serde as default feature.
+- Setters for callback functions now require `Box<dyn FnMut(...) + Send>`
+  instead of `Option<impl FnMut(...) + Send + 'static>`.
 
 ## v0.3.5
 ### API Changes